公家的物件抽出來
