+++ 臺灣言語工具/翻譯/摩西工具/安裝摩西翻譯佮相關程式.py
@@ -1,15 +1,14 @@
 # -*- coding: utf-8 -*-
-from 臺灣言語工具.系統整合.程式腳本 import 程式腳本
 from os import makedirs
 from os.path import join, isdir
 from shutil import copy
 
 
 from 臺灣言語工具.系統整合.外部程式 import 外部程式
+from 臺灣言語工具.系統整合.安裝程式腳本 import 安裝程式腳本
 
 
-class 安裝摩西翻譯佮相關程式(程式腳本):
-    pull深度 = '100'
+class 安裝摩西翻譯佮相關程式(安裝程式腳本):
 
     @classmethod
     def 安裝moses(cls, moses安裝路徑=外部程式.目錄(), 編譯CPU數=4):
@@ -67,31 +66,4 @@ class 安裝摩西翻譯佮相關程式(程式腳本):
             copy(
                 join('scripts', 'merge_alignment.py'),
                 join('bin', 'merge_alignment.py')
-            )
-
-    @classmethod
-    def _下載專案(cls, 專案git路徑):
-        try:
-            cls._走指令([
-                'git', 'clone',
-                '--depth', cls.pull深度,
-                專案git路徑
-            ])
-        except:
-            cls._走指令([
-                'git', 'clone',
-                專案git路徑
-            ])
-
-    @classmethod
-    def _更新專案(cls):
-        try:
-            cls._走指令(
-                ['git', 'pull', '--depth', cls.pull深度],
-                愛直接顯示輸出=True
-            )
-        except:
-            cls._走指令(
-                ['git', 'pull'],
-                愛直接顯示輸出=True
             )
