+++ 臺灣言語工具/語音辨識/HTK工具/安裝HTK語音辨識程式.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from 臺灣言語工具.系統整合.程式腳本 import 程式腳本
+from 臺灣言語工具.系統整合.安裝程式腳本 import 安裝程式腳本
 from os import makedirs
 from os.path import join, isdir
 
@@ -7,8 +7,7 @@ from os.path import join, isdir
 from 臺灣言語工具.系統整合.外部程式 import 外部程式
 
 
-class 安裝HTK語音辨識程式(程式腳本):
-    pull深度 = '100'
+class 安裝HTK語音辨識程式(安裝程式腳本):
 
     @classmethod
     def 安裝htk(cls, htk安裝路徑=外部程式.目錄()):
@@ -16,29 +15,12 @@ class 安裝HTK語音辨識程式(程式腳本):
         htk程式碼目錄 = join(htk安裝路徑, 'HTK_HTS')
         if not isdir(htk程式碼目錄):
             with cls._換目錄(htk安裝路徑):
-                try:
-                    cls._走指令([
-                        'git', 'clone',
-                        '--depth', cls.pull深度,
-                        'https://github.com/a8568730/HTK_HTS.git'
-                    ])
-                except:
-                    cls._走指令([
-                        'git', 'clone',
-                        'https://github.com/a8568730/HTK_HTS.git'
-                    ])
+                cls._下載專案(
+                    'https://github.com/a8568730/HTK_HTS.git'
+                )
         else:
             with cls._換目錄(htk程式碼目錄):
-                try:
-                    cls._走指令(
-                        ['git', 'pull', '--depth', cls.pull深度],
-                        愛直接顯示輸出=True
-                    )
-                except:
-                    cls._走指令(
-                        ['git', 'pull'],
-                        愛直接顯示輸出=True
-                    )
+                cls._更新專案()
         with cls._換目錄(htk程式碼目錄):
             try:
                 cls._走指令(['make', 'all', 'install'])
